Update codespell config and minor corrections (#816)
# References and relevant issues

Make more consistent with Scientific Python recommended practice:
https://learn.scientific-python.org/development/guides/style/#spelling

# Description

This PR simplifies the codespell configuration to be more explicit. It
adds `make spelling` option for easier testing in development. It adds
an `ignore-words.txt` since it is more readable than the regex in the
current config. It skips some additional files and folders.