Reorganize bundle instructions page to make it easier to navigate and provide download links (#813)
Closes: https://github.com/napari/docs/issues/714
Note: this is a reorganization borne out of a review by Harry Barng at
SciPy 2025

# Description
This PR makes it simpler to navigate the bundle installation:
1. adds tiles to jump to the instructions for a chosen OS
2. adds direct links to the proper GitHub download for the version. Note
I tried to use tiles for this, but the version substitution didn't work.
I'm open to other suggestions for how to do it. E.g. it might be
possible to use an image as the click target. After discussing with
Harry we decided that best was the enemy of the good and this was still
worth implementing.

I also made some copy edits, mostly to just standardize on "app" for
this, since bundle isn't at all obvious to the target audience, while
everyone is familiar with apps.