Add draft of 0.6.5 release notes (#845)
# References and relevant issues

#835

# Description

When I try to see the problem with @jaimergp name in the release notes,
I try to reproduce it and run the script. I fail to reproduce it.

I have used 
```bash
python generate_release_notes.py 0.6.5 --target-directory ../napari-docs/docs/release
```

As it is generated, I create a PR to reduce the risk of failing with the
alpha release.

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>