Add funding information to the home page (#818)
This PR does three things:

- adds a "funding" section to the home page, which we aim to use for
  fundraising
- update out-of-date numbers like the numbers of contributors, and
changes the
language slightly to reflect that Talley is no longer actively
contributing.
- updates the IFP section in the about page, to include recent CZI
grants and
matching funds, and grants from NumFOCUS. It also clarifies what the
IFPs are
  by linking to the governance.

This PR does not do:

- bring the project history up to date (#584). That's a longer lift, and
it
  needs to be done, but I'd like to keep that in a separate PR.
- add details about the scaffolding grant, which I felt was hard to do
  concisely. We need a blog post about it.

I am struggling with alignment of the logos, so if anyone has
suggestions for
those, please actually just push to this branch. 😂

---------

Co-authored-by: Tim Monko <timmonko@gmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>