Update and restructure the preferences guide (#834)
# References and relevant issues
Closes https://github.com/napari/docs/issues/629
Closes https://github.com/napari/docs/issues/681
Closes https://github.com/napari/docs/issues/13
Zulip comment from Grzegorz about overriding settings globally:

https://napari.zulipchat.com/#narrow/channel/212875-general/topic/.E2.9C.94.20global.20settings/near/537880501

# Description
In this PR I update and restructure the Preferences guide to be a bit
more informative.
1. I add information about the fact that preferences are per env and
where preferences are stored
2. I move the individual setting items to dropdowns in each settings
section, under the screenshots
3. I add information on overriding the settings using env variables --
in doing this I opted to add the env vars to the details for each
setting. Here I include info about overriding at the CLI/script, but
also at the shell/shellrc level, as per the comment from Grzegorz.

Additional changes: 
- there were some settings presented that are internal not intended to
be modified by users. I've simply hidden them.
- The Shortcuts section was horrible. I was going to remove it, but with
a bit of help from Claude got it to be formatted in a table. This could
be improved, e.g. by getting the shortcuts to render as keys, but for
now I think it's a strict improvement.
- I also added links to this guide in a few places including using a
`tip` admonition in the viewer guide.


@napari-bot make html-noplot

---------

Co-authored-by: Melissa Weber Mendonça <melissawm@gmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Tim Monko <timmonko@gmail.com>