Proposed roadmap updates for Q3 (#873)
# Description
As discussed at core team meeting and in various other places, this PR
makes a few small updates to the roadmap based on our work and thoughts
from Q3. Please weigh in if you'd like anything else updated!

jni I know you were waiting for this PR so here it is!

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Tim Monko <timmonko@gmail.com>